update for update deamon

+++ install.sh
@@ -29,15 +29,25 @@ fi
 echo ">>> Creating fNIRS Network Service"
 mv ./resources/networkManager/networkHandler.py /root/
 mv ./resources/networkManager/portalAddress /root/
-mv ./resources/updateDaemon/updateDaemon.py /root/
 mv ./resources/OLED/oledlib.py /root/
 mv ./resources/OLED/logoBitmap.png /root/
 chmod 777 /root/networkHandler.py
 chmod 777 /root/oledlib.py
-chmod 777 /root/updateDaemon.py
 mv ./resources/networkManager/fnirsNetwork.service /etc/systemd/system/
 systemctl enable fnirsNetwork.service
 systemctl restart fnirsNetwork.service
+
+if [ ! -e "/root/updateDaemon.py" ]; then
+	echo ">>> Creating Update Daemon"
+	mv ./resources/updateDaemon/updateDaemon.py /root/
+	chmod 777 /root/updateDaemon.py
+fi
+if [ ! -e "/etc/systemd/system/updateDaemon.service" ]; then
+	echo ">>> Creating Update Daemon Service"
+	mv ./resources/updateDaemon/updateDaemon.service /etc/systemd/system/
+	systemctl enable updateDaemon.service
+	systemctl restart updateDaemon.servce
+fi
 
 echo ">> Cleanup"
 rm -r ./resources
